backend/requirements.txt
# Core Framework
fastapi[all]==0.115.13
uvicorn[standard]==0.34.3
starlette==0.46.2

# Database
sqlalchemy==2.0.41
asyncpg==0.30.0
alembic==1.16.2
greenlet==3.2.3

# Authentication & Security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
cryptography==45.0.4
bcrypt==4.3.0

# Data Validation & Settings
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydantic-extra-types==2.10.5
email-validator==2.2.0

# File Handling
aiofiles==24.1.0
pillow==11.2.1

# HTTP Client
httpx==0.28.1
requests==2.32.4

# JSON Processing (choose one)
orjson==3.10.18

# Logging & Monitoring
structlog==25.4.0

# GraphQL (Optional)
strawberry-graphql==0.275.2
graphql-core==3.2.6

# Environment & Config
python-dotenv==1.1.1
pyyaml==6.0.2

# System Monitoring
psutil==7.0.0

# Utilities
click==8.2.1
rich==14.0.0
